refactor: extract PaperTradingPage data logic into usePaperTradingData hook (F5)
Move the portfolio/options fetch, state, and trade handlers into a
page-local hook (components/paper/usePaperTradingData.js). PaperTradingPage
becomes a render shell that calls the hook, applies its loading guard, and
renders. The `if (loading && !portfolio) return <Syncing/>` guard stays in
the component (a hook must return data, not JSX) — caught by the page's
test suite and placed correctly.

Behavior-preserving: hook body moved verbatim (same hook order/effect deps),
returning exactly the state and handlers the JSX + guard consume. Verified
by the existing 2-test suite, full suite (83), lint, and build.
